+++ onediff_comfy_nodes/_nodes.py
@@ -26,7 +26,7 @@ from .utils import (
     load_graph,
     OUTPUT_FOLDER,
 )
-from .utils.model_patcher import state_dict_hook
+from .utils.model_patcher import state_dict_hook, get_mixed_speedup_class
 from .utils.loader_sample_tools import compoile_unet, quantize_unet
 from .utils.graph_path import generate_graph_path
 from .modules.hijack_model_management import model_management_hijacker
@@ -48,53 +48,6 @@ __all__ = [
 
 if not args.dont_upcast_attention:
     os.environ["ONEFLOW_KERENL_FMHA_ENABLE_TRT_FLASH_ATTN_IMPL"] = "0"
-
-def get_mixed_speedup_class(module_cls):
-    class MixedSpeedUpModelPatcher(OneFlowSpeedUpModelPatcher, module_cls):
-        def __init__(
-            self,
-            model_patcher,
-            load_device,
-            offload_device,
-            size=0,
-            current_device=None,
-            weight_inplace_update=False,
-            *,
-            use_graph=None,
-        ):
-            self.__dict__.update(**model_patcher.__dict__)
-            OneFlowSpeedUpModelPatcher.__init__(
-                self,
-                model_patcher.model,
-                load_device,
-                offload_device,
-                size,
-                current_device,
-                weight_inplace_update,
-                use_graph=use_graph,
-            )
-
-        def clone(self):
-            cloned = module_cls.clone(self)
-            n = OneFlowSpeedUpModelPatcher(
-                cloned.model,
-                self.load_device,
-                self._1f_offload_device,
-                self._1f_size,
-                self._1f_current_device,
-                weight_inplace_update=self._1f_weight_inplace_update,
-            )
-            n.patches = {}
-            for k in self.patches:
-                n.patches[k] = self.patches[k][:]
-
-            n.object_patches = self.object_patches.copy()
-            n.model_options = copy.deepcopy(self.model_options)
-            n.model_keys = self.model_keys
-            return n
-
-
-    return MixedSpeedUpModelPatcher
 
 
 class ModelSpeedup:
